EES-1815 Move methodologyFile.ToFileInfo to MethodologyFileExtensions
