+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/Extensions/MethodologyFileExtensions.cs
@@ -1,4 +1,6 @@
-﻿namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Extensions
+﻿using GovUk.Education.ExploreEducationStatistics.Common.Model;
+
+namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Extensions
 {
     public static class MethodologyFileExtensions
     {
@@ -6,5 +8,18 @@
         {
             return methodologyFile.File.Path();
         }
+
+        public static FileInfo ToFileInfo(this MethodologyFile methodologyFile, BlobInfo blobInfo)
+        {
+            return new FileInfo
+            {
+                Id = methodologyFile.File.Id,
+                FileName = methodologyFile.File.Filename,
+                Name = null,
+                Size = blobInfo.Size,
+                Type = methodologyFile.File.Type
+            };
+        }
+
     }
 }
